w25-scraping/pyproject.toml
[tool.poetry]
name = "w25-scraping"
version = "0.1.0"
description = ""
authors = ["nalin <74811063+lordspline@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
scrapybara = "^2.1.2"
python-dotenv = "^1.0.1"
httpx = "0.27"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
